Move CaliburnMicro project composition classes to new caliburn micro project template
