+++ templates/Uwp/Projects/Default.CaliburnMicro/.template.config/template.json
@@ -96,6 +96,17 @@
         "projectPath": "Param_ProjectName.csproj"
       },
       "continueOnError": "true"
+    },
+    {
+      "description": "Add nuget package",
+      "manualInstructions": [ ],
+      "actionId": "0B814718-16A3-4F7F-89F1-69C0F9170EAD",
+      "args": {
+        "packageId" : "Caliburn.Micro",
+        "version" : "3.2.0",
+        "projectPath": "Param_ProjectName.csproj"
+      },
+      "continueOnError": "true"
     }
   ]
 }
