+++ templates/Uwp/Projects/Default.CaliburnMicro/App.xaml
@@ -1,9 +1,9 @@
-﻿<Application
+﻿<cm:CaliburnApplication
     x:Class="Param_RootNamespace.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    >
-
+    xmlns:cm="using:Caliburn.Micro"
+    RequestedTheme="Light">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
@@ -15,4 +15,5 @@
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </Application.Resources>
-</Application>
+</cm:CaliburnApplication>
+
