+++ templates/Uwp/Projects/Default.CaliburnMicro/App.xaml.cs
@@ -1,12 +1,14 @@
 ﻿using System;
 using Windows.ApplicationModel.Activation;
 using Windows.UI.Xaml;
-
+using System.Collections.Generic;
+using Caliburn.Micro;
 using Param_RootNamespace.Services;
 
 namespace Param_RootNamespace
 {
-    public sealed partial class App : Application
+    [Windows.UI.Xaml.Data.Bindable]
+    public sealed partial class App
     {
         private Lazy<ActivationService> _activationService;
 
@@ -18,6 +20,8 @@ namespace Param_RootNamespace
         public App()
         {
             InitializeComponent();
+
+            Initialize();
 
             // Deferred execution until used. Check https://msdn.microsoft.com/library/dd642331(v=vs.110).aspx for further info on Lazy<T> class.
             _activationService = new Lazy<ActivationService>(CreateActivationService);
@@ -34,6 +38,40 @@ namespace Param_RootNamespace
         protected override async void OnActivated(IActivatedEventArgs args)
         {
             await ActivationService.ActivateAsync(args);
+        }
+
+        private WinRTContainer _container;
+
+        protected override void Configure()
+        {
+            // This configures the framework to map between MainViewModel and MainPage
+            // Normally it would map between MainPageViewModel and MainPage
+            var config = new TypeMappingConfiguration
+            {
+                IncludeViewSuffixInViewModelNames = false
+            };
+
+            ViewLocator.ConfigureTypeMappings(config);
+            ViewModelLocator.ConfigureTypeMappings(config);
+
+            _container = new WinRTContainer();
+            _container.RegisterWinRTServices();
+
+        }
+
+        protected override object GetInstance(Type service, string key)
+        {
+            return _container.GetInstance(service, key);
+        }
+
+        protected override IEnumerable<object> GetAllInstances(Type service)
+        {
+            return _container.GetAllInstances(service);
+        }
+
+        protected override void BuildUp(object instance)
+        {
+            _container.BuildUp(instance);
         }
     }
 }
