+++ templates/Uwp/Projects/Default.CaliburnMicro/Helpers/PropertyChangedBaseExtensions.cs
