--- templates/Uwp/_composition/CaliburnMicro/Project/.template.config/template.json
@@ -1,50 +1,0 @@
-﻿{
-  "author": "Nigel Sampson",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CaliburnMicro.Project",
-  "tags": {
-    "language": "C#",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform" : "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionOrder": "0",
-    "wts.licenses": "[Caliburn.Micro](https://github.com/Caliburn-Micro/Caliburn.Micro/blob/master/License.txt)",
-    "wts.compositionFilter": "$framework == CaliburnMicro & wts.type == project"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [
-    { "path": ".\\Helpers\\PropertyChangedBaseExtensions.cs" },
-    { "path": ".\\Activation\\DefaultLaunchActivationHandler.cs" },
-    { "path": ".\\Services\\ActivationService.cs" },
-    { "path": ".\\App.xaml.cs" },
-    { "path": ".\\App.xaml" }
-  ],
-  "symbols": {
-    "wts.projectName": {
-      "type": "parameter",
-      "replaces": "Param_ProjectName",
-      "fileRename": "Param_ProjectName"
-    },
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    }
-  },
-  "postActions": [
-    {
-      "description": "Add nuget package",
-      "manualInstructions": [ ],
-      "actionId": "0B814718-16A3-4F7F-89F1-69C0F9170EAD",
-      "args": {
-        "packageId" : "Caliburn.Micro",
-        "version" : "3.2.0",
-        "projectPath": "Param_ProjectName.csproj"
-      },
-      "continueOnError": "true"
-    }
-  ]
-}
