--- templates/Uwp/_composition/CaliburnMicro/Project/Activation/DefaultLaunchActivationHandler.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using Caliburn.Micro;
-using Windows.ApplicationModel.Activation;
-
-namespace Param_RootNamespace.Activation
-{
-    internal class DefaultLaunchActivationHandler : ActivationHandler<LaunchActivatedEventArgs>
-    {
-        private readonly Type _navElement;
-        private readonly INavigationService _navigationService;
-
-        public DefaultLaunchActivationHandler(Type navElement, INavigationService navigationService)
-        {
-            _navElement = navElement;
-            _navigationService = navigationService;
-        }
-
-        protected override async Task HandleInternalAsync(LaunchActivatedEventArgs args)
-        {
-            // When the navigation stack isn't restored navigate to the first page,
-            // configuring the new page by passing required information as a navigation
-            // parameter
-            _navigationService.NavigateToViewModel(_navElement, args.Arguments);
-
-            await Task.CompletedTask;
-        }
-
-        protected override bool CanHandleInternal(LaunchActivatedEventArgs args)
-        {
-            // None of the ActivationHandlers has handled the app activation
-            return _navigationService.SourcePageType == null && _navElement != null;
-        }
-    }
-}
