--- templates/Uwp/_composition/CaliburnMicro/Project/App.xaml
@@ -1,18 +1,0 @@
-﻿<cm:CaliburnApplication
-    x:Class="Param_RootNamespace.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:cm="using:Caliburn.Micro"
-    RequestedTheme="Light">
-    <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="/Styles/_Colors.xaml"/>
-                <ResourceDictionary Source="/Styles/_FontSizes.xaml"/>
-                <ResourceDictionary Source="/Styles/_Thickness.xaml"/>
-
-                <ResourceDictionary Source="/Styles/TextBlock.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Application.Resources>
-</cm:CaliburnApplication>
