--- templates/Uwp/_composition/CaliburnMicro/Project/App.xaml.cs
@@ -1,77 +1,0 @@
-﻿using System;
-using Windows.ApplicationModel.Activation;
-using Windows.UI.Xaml;
-using System.Collections.Generic;
-using Caliburn.Micro;
-using Param_RootNamespace.Services;
-
-namespace Param_RootNamespace
-{
-    [Windows.UI.Xaml.Data.Bindable]
-    public sealed partial class App
-    {
-        private Lazy<ActivationService> _activationService;
-
-        private ActivationService ActivationService
-        {
-            get { return _activationService.Value; }
-        }
-
-        public App()
-        {
-            InitializeComponent();
-
-            Initialize();
-
-            // Deferred execution until used. Check https://msdn.microsoft.com/library/dd642331(v=vs.110).aspx for further info on Lazy<T> class.
-            _activationService = new Lazy<ActivationService>(CreateActivationService);
-        }
-
-        protected override async void OnLaunched(LaunchActivatedEventArgs args)
-        {
-            if (!args.PrelaunchActivated)
-            {
-                await ActivationService.ActivateAsync(args);
-            }
-        }
-
-        protected override async void OnActivated(IActivatedEventArgs args)
-        {
-            await ActivationService.ActivateAsync(args);
-        }
-
-        private WinRTContainer _container;
-
-        protected override void Configure()
-        {
-            // This configures the framework to map between MainViewModel and MainPage
-            // Normally it would map between MainPageViewModel and MainPage
-            var config = new TypeMappingConfiguration
-            {
-                IncludeViewSuffixInViewModelNames = false
-            };
-
-            ViewLocator.ConfigureTypeMappings(config);
-            ViewModelLocator.ConfigureTypeMappings(config);
-
-            _container = new WinRTContainer();
-            _container.RegisterWinRTServices();
-
-        }
-
-        protected override object GetInstance(Type service, string key)
-        {
-            return _container.GetInstance(service, key);
-        }
-
-        protected override IEnumerable<object> GetAllInstances(Type service)
-        {
-            return _container.GetAllInstances(service);
-        }
-
-        protected override void BuildUp(object instance)
-        {
-            _container.BuildUp(instance);
-        }
-    }
-}
